backend/requirements.txt
beautifulsoup4==4.12.3
fastapi==0.116.1
openai==1.97.1
pydantic==2.11.7
python-dotenv==1.0.1
requests==2.32.3
sortedcontainers==2.4.0
uvicorn[standard]==0.30.6
